+++ CSAPI/Areas/EmployerArea/Controllers/Employer_ReccomendationsForJobController.cs
@@ -1,0 +1,90 @@
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using CSAPI.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
+using CSAPI.Areas.EmployerArea.Models;
+
+namespace CSAPI.Areas.EmployerArea.Controllers
+{
+    [Area("EmployerArea")]
+    [Route("api/[controller]/")]
+    [ApiController]
+    [Authorize(Roles = "Employer")]
+    public class Employer_ReccomendationsForJobController : ControllerBase
+    {
+        private readonly IJobsRepo _jobsRepo;
+        private readonly IEmployerAreaRepo _eRepo;
+
+        public Employer_ReccomendationsForJobController(IJobsRepo jobsRepo, IEmployerAreaRepo empRepo)
+        {
+            _jobsRepo = jobsRepo;
+            _eRepo = empRepo;
+        }
+
+        [HttpGet("RecommendationForEachJob/{id}")]
+        public async Task<List<JobSeeker>> GetRecommendations(int jobid)
+        {
+            List<JobSeeker> recommendedList = new List<JobSeeker>();
+            Tuple<JobSeeker, int> recommendedrow;
+            List<Tuple<JobSeeker, int>> recommendedListCondition = new List<Tuple<JobSeeker, int>>();
+            List<JobSeeker> IndustryRec = new List<JobSeeker>();
+            List<JobSeeker> SpecialRec = new List<JobSeeker>();
+            List<JobSeeker> SkillRec = new List<JobSeeker>();
+            List<Tuple<JobSeeker, int>> PrefSkillRec = new List<Tuple<JobSeeker, int>>();
+            int pref = 0;
+
+            //var userIdCookie = Convert.ToInt32(Request.Cookies["UserId"]);
+            var userIdClaim = User.FindFirst("UserId")?.Value;
+            bool x = int.TryParse(userIdClaim, out var userIdCookie);
+            int Empid = _eRepo.GetEmpID(userIdCookie);
+
+            var job = await _jobsRepo.FindByIdAsync(jobid);
+            if (job != null && job.EmployerID == Empid)
+            {
+                IndustryRec = await _eRepo.GetRecommendationByIndustry(jobid);
+                SpecialRec = await _eRepo.GetRecommendationBySpecialization(jobid);
+                PrefSkillRec = await _eRepo.GetRecommendationBySkills(jobid);
+                SkillRec = PrefSkillRec.Select(x => x.Item1).ToList();
+
+                foreach (var i in SkillRec)
+                {
+                    pref = 0;
+                    if (IndustryRec.Contains(i) && SpecialRec.Contains(i))
+                        pref = 1;
+                    else if (IndustryRec.Contains(i) || SpecialRec.Contains(i))
+                        pref = 2;
+                    else pref = 3;
+                    recommendedrow = new Tuple<JobSeeker, int>(i, pref);
+                    recommendedListCondition.Add(recommendedrow);
+                }
+                foreach (var i in SpecialRec)
+                {
+                    if (IndustryRec.Contains(i))
+                        pref = 2;
+                    else pref = 3;
+                    if (!SkillRec.Contains(i))
+                    {
+                        recommendedrow = new Tuple<JobSeeker, int>(i, pref);
+                        recommendedListCondition.Add(recommendedrow);
+                    }
+                }
+                foreach (var i in IndustryRec)
+                {
+                    if (!SkillRec.Contains(i) && !SpecialRec.Contains(i))
+                    {
+                        pref = 3;
+                        recommendedrow = new Tuple<JobSeeker, int>(i, pref);
+                        recommendedListCondition.Add(recommendedrow);
+                    }
+                }
+
+                recommendedList = recommendedListCondition.Select(x => x.Item1).ToList();
+                return await Task.FromResult(recommendedList);
+            }
+            else
+                return null;
+        }
+
+    }
+}
